pyproject.toml
[tool.poetry]
name = "ml_pipeline_localStack"

version = "0.0.1"
authors = ["Andrey Khaykin <andreykhaykin@gmail.com>"]
license = "public"
description = ""

[tool.poetry.dependencies]
python = "^3.7"
tqdm = "^4.61.0"
boto3 = "1.17.100"
Pillow = "^8.3.0"
albumentations = "^1.0.0"


[tool.poetry.dev-dependencies]
pre-commit = "^2.7.1"
black = "^20.8b1"
pyupdate = "^1.4.0"
pyupgrade = "^1.1.0"
isort = "^5.6.4"
flake8 = "^3.8.4"

[tool.black]
# https://github.com/psf/black
line-length = 88
target-version = ["py37"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
# https://github.com/timothycrosley/isort/
include_trailing_comma = true
multi_line_output = 3
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
